requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.2
django-autoslug==1.9.8
django-crispy-forms==1.14.0
django-filter==21.1
django-qr-code==3.0.0
djangorestframework==3.13.1
frozenlist==1.3.0
geoip2==4.5.0
idna==3.3
maxminddb==2.2.0
multidict==6.0.2
Pillow==9.0.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
segno==1.4.1
sqlparse==0.4.2
twilio==7.5.1
tzdata==2021.5
urllib3==1.26.8
yarl==1.7.2
